Add a postcode validator
The validator sanitizes the postcode on initialization to ensure
a consistent postcode format to work with. Once cleaned it begins
validations.

First it checks there are only alphanumeric characters and is not
an empty string. Then it checks to see if the postcode is overridden
this is to prevent an overhead of an API request unneedlessly

If all the above passes through we do a postcode lookup for the LSOA
from the postcode.io service, validate the possible error cases first
then validates if the LSOA of the postcode is supported.

Returns both if the lsoa is supported and the LSOA returned from the API